requirements.txt
opencv-python
mediapipe==0.10.20
numpy
Pillow
gtts
playsound
pyserial